Use useAnswerReference hook in both reference controls
