+++ src/components/BaseQuestionnaireResponseForm/widgets/reference.tsx
@@ -4,15 +4,22 @@ import _ from 'lodash';
 import { ActionMeta, MultiValue, SingleValue } from 'react-select';
 import { parseFhirQueryExpression, QuestionItemProps } from 'sdc-qrf';
 
-import { QuestionnaireItemAnswerOption, QuestionnaireResponseItemAnswer, Resource } from '@beda.software/aidbox-types';
+import {
+    QuestionnaireItemAnswerOption,
+    QuestionnaireResponseItemAnswer,
+    QuestionnaireResponseItemAnswerValue,
+    Resource,
+} from '@beda.software/aidbox-types';
 import { ResourcesMap } from '@beda.software/fhir-react';
 import { buildQueryParams, isSuccess } from '@beda.software/remote-data';
 
 import { AsyncSelect } from 'src/components/Select';
-import { loadResourceOptions } from 'src/services/questionnaire';
-import { getAnswerCode, getAnswerDisplay } from 'src/utils/questionnaire';
+import { LoadResourceOption, loadResourceOptions } from 'src/services/questionnaire';
+import { getAnswerCode, getAnswerDisplay, getDisplay } from 'src/utils/questionnaire';
 
 import { useFieldController } from '../hooks';
+import { useCallback, useEffect, useState } from 'react';
+import { getDateTypeDisplay } from './utils';
 
 export type AnswerReferenceProps<R extends Resource, IR extends Resource> = QuestionItemProps & {
     overrideGetDisplay?: (resource: R, includedResources: ResourcesMap<R | IR>) => string;
@@ -95,7 +102,8 @@ export function useAnswerReference<R extends Resource = any, IR extends Resource
     context,
     overrideGetDisplay,
 }: AnswerReferenceProps<R, IR>) {
-    const { linkId, repeats, required, answerExpression, choiceColumn, text, entryFormat } = questionItem;
+    const { linkId, repeats, required, answerExpression, choiceColumn, text, entryFormat, renderingStyle, dataType } =
+        questionItem;
     const rootFieldPath = [...parentPath, linkId];
     const fieldPath = [...rootFieldPath, ...(repeats ? [] : ['0'])];
     const rootFieldName = rootFieldPath.join('.');
@@ -103,25 +111,36 @@ export function useAnswerReference<R extends Resource = any, IR extends Resource
     const fieldName = fieldPath.join('.');
     const fieldController = useFieldController(fieldPath, questionItem);
 
-    const getDisplay =
-        overrideGetDisplay ?? ((resource: R) => fhirpath.evaluate(resource, choiceColumn![0]!.path!, context)[0]);
+    const getValueDisplay = (value: QuestionnaireResponseItemAnswerValue) => {
+        if (!renderingStyle || !dataType) {
+            return getDisplay(value);
+        }
+
+        return getDateTypeDisplay({ dataType, renderingStyle, value: value.Reference?.display });
+    };
 
     // TODO: add support for fhirpath and application/x-fhir-query
     const [resourceType, searchParams] = parseFhirQueryExpression(answerExpression!.expression!, context);
 
-    const loadOptions = async (searchText: string) => {
-        const response = await loadResourceOptions(
-            resourceType as any,
-            { ...(typeof searchParams === 'string' ? {} : searchParams ?? {}), _ilike: searchText },
-            getDisplay,
-        );
+    const [loadedOptions, setLoadedOptions] = useState<LoadResourceOption<R>[] | null>(null);
 
-        if (isSuccess(response)) {
-            return response.data;
-        }
+    const loadOptions = useCallback(
+        async (searchText: string) => {
+            const response = await loadResourceOptions(
+                resourceType as any,
+                { ...(typeof searchParams === 'string' ? {} : searchParams ?? {}), _ilike: searchText },
+                overrideGetDisplay ??
+                    ((resource: R) => fhirpath.evaluate(resource, choiceColumn![0]!.path!, context)[0]),
+            );
 
-        return [];
-    };
+            if (isSuccess(response)) {
+                return response.data;
+            }
+
+            return [];
+        },
+        [resourceType, searchParams, choiceColumn, context, overrideGetDisplay],
+    );
 
     const debouncedLoadOptions = _.debounce(
         (searchText: string, callback: (options: QuestionnaireItemAnswerOption[]) => void) => {
@@ -129,6 +148,17 @@ export function useAnswerReference<R extends Resource = any, IR extends Resource
         },
         500,
     );
+
+    useEffect(() => {
+        if (questionItem.itemControl) {
+            const loadItemControlOptions = async () => {
+                const options = await loadOptions('');
+                setLoadedOptions(options);
+            };
+
+            loadItemControlOptions();
+        }
+    }, [questionItem.itemControl]);
 
     const onChange = (
         _value: SingleValue<QuestionnaireItemAnswerOption> | MultiValue<QuestionnaireItemAnswerOption>,
@@ -163,6 +193,8 @@ export function useAnswerReference<R extends Resource = any, IR extends Resource
         rootFieldName,
         fieldName,
         debouncedLoadOptions,
+        loadedOptions,
+        getValueDisplay,
         onChange,
         validate,
         searchParams,
@@ -175,18 +207,14 @@ export function useAnswerReference<R extends Resource = any, IR extends Resource
     };
 }
 
-// function buildRules(props: QuestionnaireItem) {
-//     const { required } = props;
-//     if (required) {
-//         return [{ required: true, message: 'This field is required' }];
-//     }
-//     return [];
-// }
-
 function QuestionReferenceUnsafe<R extends Resource = any, IR extends Resource = any>(
     props: AnswerReferenceProps<R, IR>,
 ) {
-    const { debouncedLoadOptions, fieldController, repeats, placeholder } = useAnswerReference(props);
+    const { questionItem } = props;
+    const { repeats } = questionItem;
+
+    const { debouncedLoadOptions, fieldController, placeholder } = useAnswerReference(props);
+
     const { formItem } = fieldController;
 
     return (
